animation/requirements.txt
rasterio
pandas
geopandas
matplotlib
geoplot
pillow
joblib
